+++ backend/src/main/java/com/ssafy/togeball/domain/matching/repository/MatchingRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.togeball.domain.matching.repository;
 import com.ssafy.togeball.domain.matching.entity.Matching;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchingRepository extends JpaRepository<Matching, Long> {
+public interface MatchingRepository extends JpaRepository<Matching, Integer> {
 
 }
