+++ backend/src/main/java/com/ssafy/togeball/domain/matching/repository/MatchingTagRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.togeball.domain.matching.repository;
 import com.ssafy.togeball.domain.matching.entity.MatchingTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchingTagRepository extends JpaRepository<MatchingTag, Long> {
+public interface MatchingTagRepository extends JpaRepository<MatchingTag, Integer> {
 }
