+++ backend/src/main/java/com/ssafy/togeball/domain/notice/repository/NoticeRepository.java
@@ -8,11 +8,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
+public interface NoticeRepository extends JpaRepository<Notice, Integer> {
 
 //    Optional<Notice> findByMatchingIdAndUserId(Long matchingId, Long userId);
 
-    List<Notice> findByUserId(Long userId);
+    List<Notice> findByUserId(Integer userId);
 
 //    List<Notice> findByMatchingId(Long matchingId);
 }
